Rename common.py to test_common.py
This makes the file be included on tarballs by the build system.

Fixes #170
