+++ test/test_client_rpc.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from nose.tools import with_setup, eq_ as eq
-from common import vim, cleanup
+from test_common import vim, cleanup
 
 cid = vim.channel_id
