+++ test/test_events.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from nose.tools import with_setup, eq_ as eq
-from common import vim, cleanup
+from test_common import vim, cleanup
 
 
 @with_setup(setup=cleanup)
